fix: Fix rendering index in dev
