+++ site/routes.json
@@ -1,14 +1,14 @@
 {
   "meta": {
     "language": "en",
-    "siteName": "Gustwind",
-    "scripts": [
-      {
-        "type": "text/javascript",
-        "src": "https://unpkg.com/sidewind@6.0.0/dist/sidewind.umd.production.min.js"
-      }
-    ]
+    "siteName": "Gustwind"
   },
+  "scripts": [
+    {
+      "type": "text/javascript",
+      "src": "https://unpkg.com/sidewind@6.0.0/dist/sidewind.umd.production.min.js"
+    }
+  ],
   "routes": {
     "/": {
       "layout": "siteIndex",
